refactor: biome fixes
